added passport, passport-local, express, and express-session as dependencies(might be one more but i can't remember).
Created the account page.
user's Account Page will show information appropriately to each user.
Deleted a single semi-colon from listing page because it was causing a warning to appear.
Edited backend, loginpage, app.jsx, and header to be able to authenticate users on login and allow users to signout after logging in.
Edited Header so that if a user is logged in, the login option will hide and the 'MyAccount' and 'Signout' options will appear.
Added a modal in header so that the user is notified when trying to logout and double checks with user if they want to logout.
Added a notification to let user know when they log in that they are logged in.

added an auth-cookies file in case we ever want to use encryption for session details(currently no encryption because i could not get it to work properly).
deleted css related to account page because i used className instead.

changed useEffect in app.jsx to properly handle checking login.
Created About Page but nothing on page yet.